add some alogrim when delete category, product set to default category

+++ usecase/category.js
@@ -102,10 +102,40 @@ class CategoryUC {
       result.status = 404;
       return result;
     }
+    if (category.name === "other") {
+      result.reason = 'cant not delet default category'
+      result.status = 400
+      return result
+    }
+    let product = await this.productRepository.getProductByCategoryId(id)
+    await this.setDefaultCategory(product)
     await this.categoryRepository.deleteCategory(id);
     result.isSuccess = true;
     result.status = 200;
     return result;
+  }
+
+  async setDefaultCategory(products) {
+    for (let i = 0; i < products.length; i++) {
+
+      let defaultCategory = await this.categoryRepository.getDefaultCategory("other")
+      if (defaultCategory === null) {
+        let dataCreate = {
+          name: "other"
+        }
+        let newCategory = await this.categoryRepository.addCategory(dataCreate)
+        let newProduct = {
+          category_id: newCategory.id
+        }
+        await this.productRepository.updateProduct(newProduct, products[i].id)
+      } else {
+        let updateData = {
+          category_id: defaultCategory.id
+        }
+        await this.productRepository.updateProduct(updateData, products[i].id)
+      }
+
+    }
   }
 }
